+++ src/main/java/techcourse/myblog/domain/article/Contents.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.article;
+package techcourse.myblog.domain.article;
 
 import javax.persistence.Embeddable;
 import javax.persistence.Lob;
