+++ src/main/java/techcourse/myblog/domain/comment/Contents.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.comment;
+package techcourse.myblog.domain.comment;
 
 import javax.persistence.Embeddable;
