+++ src/main/java/techcourse/myblog/domain/user/Information.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.user;
+package techcourse.myblog.domain.user;
 
 import javax.persistence.Embeddable;
 import javax.validation.constraints.Email;
