+++ src/main/java/techcourse/myblog/domain/user/User.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.user;
+package techcourse.myblog.domain.user;
 
 import javax.persistence.*;
