+++ src/main/java/techcourse/myblog/dto/request/UserChangeableInfoDto.java
@@ -4,8 +4,8 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
 
 import org.hibernate.validator.constraints.Length;
-import techcourse.myblog.user.Information;
-import techcourse.myblog.user.User;
+import techcourse.myblog.domain.user.Information;
+import techcourse.myblog.domain.user.User;
 
 public class UserChangeableInfoDto {
 	private static final String BLANK_NAME = "이름을 입력해주세요.";
