+++ src/main/java/techcourse/myblog/repository/UserRepository.java
@@ -3,7 +3,7 @@ package techcourse.myblog.repository;
 import java.util.List;
 import java.util.Optional;
 
-import techcourse.myblog.user.User;
+import techcourse.myblog.domain.user.User;
 
 import org.springframework.data.jpa.repository.JpaRepository;
