+++ src/main/java/techcourse/myblog/service/CommentService.java
@@ -1,12 +1,12 @@
 package techcourse.myblog.service;
 
-import techcourse.myblog.article.Article;
-import techcourse.myblog.comment.Comment;
-import techcourse.myblog.comment.Contents;
+import techcourse.myblog.domain.article.Article;
+import techcourse.myblog.domain.comment.Comment;
+import techcourse.myblog.domain.comment.Contents;
+import techcourse.myblog.domain.user.User;
 import techcourse.myblog.exception.NotFoundCommentException;
 import techcourse.myblog.exception.UnauthorizedException;
 import techcourse.myblog.repository.CommentRepository;
-import techcourse.myblog.user.User;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
