+++ src/main/java/techcourse/myblog/service/UserService.java
@@ -2,14 +2,14 @@ package techcourse.myblog.service;
 
 import java.util.List;
 
+import techcourse.myblog.domain.user.Information;
+import techcourse.myblog.domain.user.User;
 import techcourse.myblog.dto.request.UserChangeableInfoDto;
 import techcourse.myblog.dto.request.UserSignUpInfoDto;
 import techcourse.myblog.exception.AlreadyExistEmailException;
 import techcourse.myblog.exception.NotFoundUserException;
 import techcourse.myblog.exception.NotMatchPasswordException;
 import techcourse.myblog.repository.UserRepository;
-import techcourse.myblog.user.Information;
-import techcourse.myblog.user.User;
 
 import org.springframework.stereotype.Service;
