+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -5,10 +5,10 @@ import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import techcourse.myblog.custom.LoginUser;
+import techcourse.myblog.domain.user.User;
 import techcourse.myblog.dto.request.UserChangeableInfoDto;
 import techcourse.myblog.dto.request.UserSignUpInfoDto;
 import techcourse.myblog.service.UserService;
-import techcourse.myblog.user.User;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
