+++ src/test/java/techcourse/myblog/repository/UserRepositoryTests.java
@@ -1,7 +1,7 @@
 package techcourse.myblog.repository;
 
 import org.junit.jupiter.api.Test;
-import techcourse.myblog.user.User;
+import techcourse.myblog.domain.user.User;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
